requirements.txt
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
PyYAML~=6.0